management/server/requirements-light.txt
Flask==3.0.3
flask_cors==5.0.0
mysql-connector-python==9.2.0
pycryptodomex==3.20.0
tabulate==0.9.0
Werkzeug==3.1.3
PyJWT==2.10.1
dotenv==0.9.9
elasticsearch==8.12.0
minio==7.2.4
strenum==0.4.15
peewee==3.17.1
pytz==2020.5
bs4==0.0.2
openai==1.70.0
redis==6.2.0
hanziconv==0.3.2
nltk==3.9.1
openpyxl==3.1.5
xlrd==2.0.1
requests>=2.31.0
